Preceding changes

+++ src/pages/StockPage.tsx
@@ -145,7 +145,7 @@ export default function StockPage() {
       buying_price: parseFloat(form.buying_price) || 0,
       wholesale_price: parseFloat(form.wholesale_price) || 0,
       retail_price: parseFloat(form.retail_price) || 0,
-      quantity: parseInt(form.quantity) || 0,
+      quantity: parseFloat(form.quantity) || 0,
       min_stock_level: parseInt(form.min_stock_level) || 5,
       tax_rate: parseFloat(form.tax_rate) || 0,
       image_url_1: editItem?.image_url_1 || '',
